pyproject.toml
[tool.poetry]
name = "hack-uk-ai-pdf-tagger"
version = "0.1.0"
description = ""
authors = ["\"Akshay Akula\" <akulaakshay30@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
numpy = "^2.1.1"
pymupdf = "^1.24.11"
pymupdf4llm = "^0.0.17"
python-dotenv = "^1.0.1"
graphviz = "^0.20.1"
pygraphviz = "^1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
